requirements-prod.txt
fastapi==0.115.8
uvicorn==0.34.0
gunicorn==21.2.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.2
supabase==2.10.0
openai==1.63.0
httpx==0.27.0
pydantic==2.4.2
email-validator==2.0.0.post2
python-multipart==0.0.6
playwright==1.53.0
cryptography==45.0.5